Deformable-DETR-surgical_tool/requirements_surgical_tool.txt
absl-py==0.15.0
adversarial-robustness-toolbox==1.9.1
antlr4-python3-runtime==4.8
anyio==3.5.0
apache-beam==2.22.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
armory-testbed==0.14.3
art==5.4
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
audioread==2.1.9
avro-python3==1.9.2.1
Babel==2.9.1
backcall==0.2.0
black==21.7b0
bleach==4.1.0
boto3==1.17.20
botocore==1.20.112
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
coloredlogs==15.0.1
colour==0.1.5
colour-science==0.4.1
crcmod==1.7
cycler==0.11.0
dataclasses==0.6
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
docker==5.0.3
docopt==0.6.2
entrypoints==0.3
executing==0.8.2
fastavro==0.23.6
ffmpeg-python==0.2.0
flatbuffers==1.12
fonttools==4.28.5
future==0.18.2
gast==0.3.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.54.0
grpcio==1.32.0
h5py==2.10.0
hdfs==2.6.0
httplib2==0.17.4
humanfriendly==10.0
hydra-core==1.0.7
idna==3.3
imageio==2.16.1
importlib-metadata==4.10.1
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.1
jupyter-core==4.9.1
jupyter-packaging==0.7.12
jupyter-server==1.13.3
jupyterlab==3.0.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
librosa==0.8.1
llvmlite==0.36.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mean-average-precision==2021.4.26.0
mistune==0.8.4
mock==2.0.0
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.8
notebook==6.4.7
numba==0.53.1
numpy==1.22.2
oauth2client==3.0.0
oauthlib==3.1.1
omegaconf==2.0.6
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==21.3
pandas==1.2.4
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pluggy==0.13.1
pooch==1.5.2
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.24
protobuf==3.19.3
ptyprocess==0.7.0
pure-eval==0.2.1
py==1.11.0
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
pydot==1.4.2
pydub==0.24.1
Pygments==2.11.2
pymongo==3.12.3
pyparsing==3.0.6
pyrsistent==0.18.1
pytest==6.2.2
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.8
s3transfer==0.3.7
scikit-image==0.19.2
scikit-learn==0.24.2
scipy==1.7.3
Send2Trash==1.8.0
six==1.15.0
sniffio==1.2.0
SoundFile==0.10.3.post1
sox==1.4.1
stack-data==0.1.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==3.2.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
tensorflow-metadata==1.6.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2022.4.8
toml==0.10.2
tomli==1.2.3
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.12.1
zipp==3.7.0
